Put the matcher into a class
This way, we don't have to pass the list of regexes each time we call
the function, which slightly increases performance, but more importantly
simplifies usage.

I've also removed the item "unspaghettify" from the TODO, because after
reading it for the second time, I've felt that the code is pretty clear
and any attempts at making it more readable would make it slower.